more docstring for synbody_utils
